% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acti_plot_summary.R
\name{acti_plot_actogram}
\alias{acti_plot_actogram}
\title{Plot an activity actogram}
\usage{
acti_plot_actogram(data, value, time = time, breaks = "4 hours", ...)
}
\arguments{
\item{data}{Minute-level activity data.}

\item{value}{Activity column, supplied unquoted or as a string.}

\item{time}{Timestamp column.}

\item{breaks}{Time-of-day axis break interval.}

\item{...}{Arguments passed to ggplot2::geom_tile().}
}
\value{
A ggplot object.
}
\description{
Displays minute-level activity as a date-by-time heat map using actibase day
indexing and time-of-day conversion.
}
