% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actiplot-package.R
\docType{package}
\name{actiplot-package}
\alias{actiplot}
\alias{actiplot-package}
\title{actiplot: Plotting Minute-Level Activity Data}
\description{
Provides 'ggplot2' plots for minute-level activity data, including step counts and activity counts. Time series can be displayed over the full recording period or aligned by time of day and faceted into one row per date. Includes example activity data described by Chadwell et al. (2019) \doi{10.1038/s41597-019-0211-6}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jhuwit/actiplot}
  \item \url{https://jhuwit.github.io/actiplot/}
  \item Report bugs at \url{https://github.com/jhuwit/actiplot/issues}
}

}
\author{
\strong{Maintainer}: John Muschelli \email{muschellij2@gmail.com} (\href{https://orcid.org/0000-0001-6469-1750}{ORCID})

Authors:
\itemize{
  \item John Muschelli \email{muschellij2@gmail.com} (\href{https://orcid.org/0000-0001-6469-1750}{ORCID})
}

}
\keyword{internal}
