% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acti_plot_summary.R
\name{geom_activity_rug}
\alias{geom_activity_rug}
\title{Add activity marks along a datetime axis}
\usage{
geom_activity_rug(data, time = "time", ...)
}
\arguments{
\item{data}{Data containing a timestamp column.}

\item{time}{Timestamp column name.}

\item{...}{Arguments passed to ggplot2::geom_rug().}
}
\value{
A ggplot2 layer.
}
\description{
Add activity marks along a datetime axis
}
