% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_intervals.R
\name{geom_day_boundary}
\alias{geom_day_boundary}
\title{Add conventional daily boundaries}
\usage{
geom_day_boundary(boundary = c("midnight", "noon"), ...)
}
\arguments{
\item{boundary}{One of midnight or noon.}

\item{...}{Arguments passed to geom_time_vline().}
}
\value{
A ggplot2 layer.
}
\description{
Add conventional daily boundaries
}
