% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_intervals.R
\name{geom_episode}
\alias{geom_episode}
\title{Add timestamped episode intervals}
\usage{
geom_episode(data, start = "start", end = "end", ...)
}
\arguments{
\item{data}{An interval data frame.}

\item{start, end}{Names of POSIXct start and end columns.}

\item{...}{Arguments passed to ggplot2::geom_rect().}
}
\value{
A ggplot2 layer.
}
\description{
Draws one or more timestamped intervals behind an activity trace. The input
data must contain POSIXct start and end columns. This is the common layer
underpinning sleep, wear, and activity-bout annotations.
}
