% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_intervals.R
\name{geom_sleep_window}
\alias{geom_sleep_window}
\title{Add sleep-window intervals}
\usage{
geom_sleep_window(data, start = "start", end = "end", ...)
}
\arguments{
\item{data}{An interval data frame.}

\item{start, end}{Names of POSIXct start and end columns.}

\item{...}{Arguments passed to ggplot2::geom_rect().}
}
\value{
A ggplot2 layer.
}
\description{
Add sleep-window intervals
}
